fix: 修改接受的id为string，得自己转换

+++ repo/list/enter.go
@@ -87,5 +87,5 @@ func ListQuery[T any](model T, option Options) (list []T, count int, err error) 
 
 // 用于绑定的id列表
 type RemoveReq struct {
-	Ids []int `json:"ids"`
+	Ids []int64 `json:"ids"`
 }
